Lots of groundwork laid for more options and a favorites tree.
Can now sort between two threes, as well as lots of ffmpeg options
Can't imagine anyone would bother to add them for themselves. Unlikely
options was not added, though if you want one, make an issue.

Currently lacking, adding extra options to a parameter, these depends on user usage as much as the download option, and will need a way for the user to interact and input the options.